+++ init_ideas/init_models.py
@@ -22,88 +22,7 @@ class mob_model:
 				f = self.flux(i,j)
 				m[i][j], m[j][i] = f, f # symmetrical
 
-		return m
-
-
-class gravity(mob_model):
-	'''
-	The gravity human mobility model
-
-	'''
-
-	def __init__(self, pop, alpha, beta, gamma, K, **kwargs):
-		super().__init__(pop)
-		kwargs.setdefault('exp', False)
-		self.alpha = alpha # population i exponent
-		self.beta = beta # population j exponent
-		self.gamma = gamma # distance exponent
-		self.K = K # fitting parameter
-		self.exp = kwargs['exp'] # True if exponential decay function is used, False if power decay is used
-
-	def f(self, r):
-		'''
-		The distance decay function
-		'''
-		if self.exp:
-			return np.exp(-self.gamma*r)
-		else:
-			return r**(-self.gamma)
-
-	def flux(self, i, j):
-		'''
-		Takes the indices of two locations and returns the flux between them
-		'''
-
-		pop = self.pop
-		popi, popj = pop.popDist[i], pop.popDist[j]
-		r = disp(pop.locCoords[i], pop.locCoords[j])
-		n = self.K * ((popi**self.alpha)*(popj**self.beta))*self.f(r)
-		return n
-
-class radiation(mob_model):
-	'''
-	The normalised radiation human mobility model
-	'''
-	def flux(self, i, j):
-		'''
-		Takes the indices of two locations i, j and returns the average flux from i to j
-		'''
-		pop = self.pop
-		popi, popj = pop.popDist[i], pop.popDist[j]
-		popSij = pop.s(i, j)
-		n = (popi/(1-popi/pop.M()))*(popi*popj)/float((popi+popSij)*(popi+popSij+popj)) # TODO how do we define Ti here?
-
-		return n
-
-class opportunities(mob_model):
-	'''
-	The intervening opportunities model
-	'''
-	def __init__(self, pop, gamma):
-		super().__init__(pop)
-		self.gamma = gamma
-
-	def norm_factor(self, i, j):
-		pop = self.pop
-		to_sum = []
-		for k in range(pop.size):
-			popk = pop.popDist[k]
-			popSik = pop.s(i, k)
-			if k != j:
-				a = np.exp((-self.gamma*popSik))-(np.exp(-self.gamma*(popSik+popk)))
-				to_sum.append(a)
-		return sum(to_sum)
-
-	def flux(self, i, j):
-		'''
-		Takes the indices of two locations i, j and returns the average flux from i to j
-		'''
-		pop = self.pop
-		popi, popj = pop.popDist[i], pop.popDist[j]
-		popSij = pop.s(i, j)
-		a = np.exp((-self.gamma*popSij)-(np.exp(-self.gamma*(popSij)+popj)))
-		n = self.norm_factor(i, j)*popi*a
-		return n
+		return må
 
 # Test Data
